build: testing framework
