+++ codestyle/findbugs-exclude.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright Amazon.com Inc. or its affiliates.
+  ~ SPDX-License-Identifier: Apache-2.0
+  -->
+
+<FindBugsFilter>
+    <Match>
+        <Or>
+            <Bug pattern="DM_CONVERT_CASE"/>
+            <Bug pattern="SIC_INNER_SHOULD_BE_STATIC_ANON"/>
+            <!-- Added for JDK-11 executable -->
+            <Bug pattern="RCN_REDUNDANT_NULLCHECK_OF_NONNULL_VALUE"/>
+            <Bug pattern="RCN_REDUNDANT_NULLCHECK_WOULD_HAVE_BEEN_A_NPE"/>
+
+            <Bug pattern="EI_EXPOSE_REP"/>
+            <Bug pattern="EI_EXPOSE_REP2"/>
+        </Or>
+    </Match>
+    <Match>
+        <Or>
+            <Package name="software.amazon.awssdk.aws.greengrass"/>
+            <Package name="software.amazon.awssdk.aws.greengrass.model"/>
+            <Package name="software.amazon.awssdk.eventstreamrpc"/>
+            <Package name="software.amazon.awssdk.eventstreamrpc.model"/>
+            <Package name="software.amazon.awssdk.http.apache.internal.conn"/>
+        </Or>
+    </Match>
+</FindBugsFilter>
