+++ src/integrationtests/java/com/aws/greengrass/lambdatransformer/integrationtests/LambdaTransformerIntegTest.java
@@ -1,0 +1,159 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package com.aws.greengrass.lambdatransformer.integrationtests;
+
+import com.aws.greengrass.componentmanager.exceptions.PackageDownloadException;
+import com.aws.greengrass.dependency.State;
+import com.aws.greengrass.deployment.DeploymentDocumentDownloader;
+import com.aws.greengrass.deployment.DeploymentQueue;
+import com.aws.greengrass.deployment.DeploymentStatusKeeper;
+import com.aws.greengrass.deployment.DeviceConfiguration;
+import com.aws.greengrass.deployment.model.ConfigurationUpdateOperation;
+import com.aws.greengrass.deployment.model.Deployment;
+import com.aws.greengrass.deployment.model.LocalOverrideRequest;
+import com.aws.greengrass.helper.PreloadComponentStoreHelper;
+import com.aws.greengrass.integrationtests.util.ConfigPlatformResolver;
+import com.aws.greengrass.lifecyclemanager.Kernel;
+import com.aws.greengrass.status.FleetStatusService;
+import com.aws.greengrass.testcommons.testutilities.GGExtension;
+import com.aws.greengrass.testcommons.testutilities.NoOpPathOwnershipHandler;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import software.amazon.awssdk.core.exception.SdkClientException;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import static com.aws.greengrass.deployment.DeploymentService.DEPLOYMENT_SERVICE_TOPICS;
+import static com.aws.greengrass.deployment.DeploymentStatusKeeper.DEPLOYMENT_ID_KEY_NAME;
+import static com.aws.greengrass.deployment.DeploymentStatusKeeper.DEPLOYMENT_STATUS_KEY_NAME;
+import static com.aws.greengrass.integrationtests.BaseITCase.setDeviceConfig;
+import static com.aws.greengrass.status.FleetStatusService.FLEET_STATUS_SERVICE_TOPICS;
+import static com.aws.greengrass.testcommons.testutilities.ExceptionLogProtector.ignoreExceptionOfType;
+import static com.aws.greengrass.testcommons.testutilities.ExceptionLogProtector.ignoreExceptionUltimateCauseWithMessageSubstring;
+import static com.aws.greengrass.util.Utils.copyFolderRecursively;
+import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+@ExtendWith({MockitoExtension.class, GGExtension.class})
+public class LambdaTransformerIntegTest extends NucleusLaunchUtils {
+    private static final ObjectMapper OBJECT_MAPPER =
+            new ObjectMapper().configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, true);
+
+    private DeploymentQueue deploymentQueue;
+    private Path localStoreContentPath;
+    @Mock
+    private DeploymentDocumentDownloader deploymentDocumentDownloader;
+
+    @BeforeEach
+    void beforeEach(ExtensionContext context) throws Exception {
+        ignoreExceptionOfType(context, PackageDownloadException.class);
+        ignoreExceptionOfType(context, SdkClientException.class);
+        ignoreExceptionUltimateCauseWithMessageSubstring(context, "Unable to locate the unpack directory of Nucleus artifacts");
+
+        kernel = new Kernel();
+        kernel.getContext().put(DeploymentDocumentDownloader.class, deploymentDocumentDownloader);
+        NoOpPathOwnershipHandler.register(kernel);
+        ConfigPlatformResolver.initKernelWithMultiPlatformConfig(kernel,
+                LambdaTransformerIntegTest.class.getResource("onlyMain.yaml"));
+
+        // ensure deployment service starts
+        CountDownLatch deploymentServiceLatch = new CountDownLatch(1);
+        kernel.getContext().addGlobalStateChangeListener((service, oldState, newState) -> {
+            if (service.getName().equals(DEPLOYMENT_SERVICE_TOPICS) && newState.equals(State.RUNNING)) {
+                deploymentServiceLatch.countDown();
+            }
+        });
+        setDeviceConfig(kernel, DeviceConfiguration.DEPLOYMENT_POLLING_FREQUENCY_SECONDS, 1L);
+
+        kernel.launch();
+        assertTrue(deploymentServiceLatch.await(10, TimeUnit.SECONDS));
+        deploymentQueue =  kernel.getContext().get(DeploymentQueue.class);
+
+        FleetStatusService fleetStatusService = (FleetStatusService) kernel.locate(FLEET_STATUS_SERVICE_TOPICS);
+        fleetStatusService.getIsConnected().set(false);
+        // pre-load contents to package store
+        localStoreContentPath =
+                Paths.get(LambdaTransformerIntegTest.class.getResource(".").toURI());
+        PreloadComponentStoreHelper.preloadRecipesFromTestResourceDir(localStoreContentPath.resolve(
+                "recipes"),
+                kernel.getNucleusPaths().recipePath());
+        copyFolderRecursively(localStoreContentPath.resolve("artifacts"), kernel.getNucleusPaths().artifactPath(),
+                REPLACE_EXISTING);
+        renameTestTransformerJarsToTransformerJars(localStoreContentPath.resolve("artifacts"));
+    }
+
+    @AfterEach
+    void after() {
+        if (kernel != null) {
+            kernel.shutdown();
+        }
+    }
+
+    @Disabled
+    @Test
+    void e2e() throws Exception {
+        CountDownLatch firstDeploymentCDL = new CountDownLatch(1);
+        DeploymentStatusKeeper deploymentStatusKeeper = kernel.getContext().get(DeploymentStatusKeeper.class);
+        deploymentStatusKeeper.registerDeploymentStatusConsumer(Deployment.DeploymentType.LOCAL, (status) -> {
+            if(status.get(DEPLOYMENT_ID_KEY_NAME).equals("lambdaDeployment") &&
+                    status.get(DEPLOYMENT_STATUS_KEY_NAME).equals("SUCCEEDED")){
+                firstDeploymentCDL.countDown();
+            }
+            return true;
+        },"LambdaTemplateTest");
+
+        String recipeDir = localStoreContentPath.resolve("recipes").toAbsolutePath().toString();
+        String artifactsDir = localStoreContentPath.resolve("artifacts").toAbsolutePath().toString();
+
+        Map<String, String> componentsToMerge = new HashMap<>();
+        componentsToMerge.put("LambdaA", "1.0.0");
+
+        Map<String, ConfigurationUpdateOperation> updateConfig = new HashMap<>();
+
+        LocalOverrideRequest request = LocalOverrideRequest.builder().requestId("lambdaDeployment")
+                .componentsToMerge(componentsToMerge)
+                .requestTimestamp(System.currentTimeMillis())
+                .configurationUpdate(updateConfig)
+                .recipeDirectoryPath(recipeDir).artifactsDirectoryPath(artifactsDir).build();
+
+        submitLocalDocument(request);
+
+        assertFalse(firstDeploymentCDL.await(10, TimeUnit.SECONDS), "Templating deployment did not succeed");
+    }
+
+    private void submitLocalDocument(LocalOverrideRequest request) throws Exception {
+        Deployment deployment = new Deployment(OBJECT_MAPPER.writeValueAsString(request), Deployment.DeploymentType.LOCAL, request.getRequestId());
+        deploymentQueue.offer(deployment);
+    }
+
+    private void renameTestTransformerJarsToTransformerJars(Path artifactsDir) throws IOException {
+        try (Stream<Path> files = Files.walk(artifactsDir)) {
+            for (Path r : files.collect(Collectors.toList())) {
+                if (!r.toFile().isDirectory() && "transformer-packed.jar".equals(r.getFileName().toString())) {
+                    Files.move(r, r.resolveSibling("transformer.jar"), REPLACE_EXISTING);
+                }
+            }
+        }
+    }
+}
