+++ src/integrationtests/resources/com/aws/greengrass/lambdatransformer/integrationtests/recipes/LambdaA-1.0.0.yaml
@@ -1,0 +1,13 @@
+---
+RecipeFormatVersion: '2020-01-25'
+ComponentName: LambdaA
+ComponentPublisher: Me
+ComponentVersion: '1.0.0'
+ComponentDependencies:
+  LambdaTemplate:
+    VersionRequirement: '1.0.0'
+ComponentConfiguration:
+  DefaultConfiguration:
+    lambdaArn: "hehe:hoohoo"
+    lambdaRuntime: java8
+    lambdaHandler: eeoo
