--- src/main/java/common/models/LambdaDeviceMount.java
@@ -1,23 +1,0 @@
-package common.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import static common.Constants.DEFAULT_ADD_GROUP_OWNER;
-import static common.Constants.DEFAULT_LAMBDA_FILE_SYSTEM_PERMISSION;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class LambdaDeviceMount {
-    private String path;
-
-    @Builder.Default
-    private LambdaFilesystemPermission permission = DEFAULT_LAMBDA_FILE_SYSTEM_PERMISSION;
-
-    @Builder.Default
-    private Boolean addGroupOwner = DEFAULT_ADD_GROUP_OWNER;
-}
