--- src/main/java/common/models/LambdaFilesystemPermission.java
@@ -1,10 +1,0 @@
-package common.models;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public enum LambdaFilesystemPermission {
-    @JsonProperty("ro")
-    RO,
-    @JsonProperty("rw")
-    RW,
-}
