--- src/main/java/common/models/LambdaInputPayloadEncodingType.java
@@ -1,10 +1,0 @@
-package common.models;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public enum LambdaInputPayloadEncodingType {
-    @JsonProperty("json")
-    JSON,
-    @JsonProperty("binary")
-    BINARY,
-}
