--- src/main/java/common/models/LambdaIsolationMode.java
@@ -1,10 +1,0 @@
-package common.models;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public enum LambdaIsolationMode {
-    @JsonProperty("GreengrassContainer")
-    GREENGRASS_CONTAINER,
-    @JsonProperty("NoContainer")
-    NO_CONTAINER,
-}
