--- src/main/java/common/models/LifecycleType.java
@@ -1,12 +1,0 @@
-package common.models;
-
-public enum LifecycleType {
-    Install,
-    Run,
-    Standby,
-    Restart,
-    startup,
-    shutdown,
-    setenv
-}
-
