+++ src/test/resources/lambda/onlyMain.yaml
