--- src/test/resources/recipes/LambdaA-1.0.0.yaml
@@ -1,12 +1,0 @@
----
-RecipeFormatVersion: '2020-01-25'
-ComponentName: LambdaA
-ComponentPublisher: Me
-ComponentVersion: '1.0.0'
-ComponentDependencies:
-  LambdaTemplate:
-    VersionRequirement: '1.0.0'
-TemplateParametes:
-  lambdaArn: "hehe:hoohoo"
-  lambdaRuntime: java8
-  lambdaHandler: eeoo
