--- src/test/resources/recipes/LambdaTemplate-1.0.0.yaml
@@ -1,78 +1,0 @@
----
-recipeFormatVersion: '2020-01-25'
-componentName: LambdaTemplate
-componentType: aws.greengrass.template
-componentVersion: '1.0.0'
-componentDescription: Lambda template for local deployment
-componentPublisher: AWS Greengrass
-templateParameterSchema:
-  lambdaArn:
-    type: "string"
-    required: true
-  lambdaRuntime:
-    type: "string"
-    required: true
-  lambdaHandler:
-    type: "string"
-    required: true
-  maxIdleTimeInSeconds:
-    type: "number"
-    required: false
-    defaultValue: 60
-  lambdaEnvironmentVariables:
-    type: "object"
-    required: false
-    defaultValue: {}
-  pinned:
-    type: "boolean"
-    required: false
-    defaultValue: true
-  lambdaArgs:
-    type: "array"
-    required: false
-    defaultValue: []
-  maxInstancesCount:
-    type: "number"
-    required: false
-    defaultValue: 100
-  pubsubTopics:
-    type: "object"
-    required: false
-    defaultValue: []
-  containerParams:
-    type: "object"
-    required: false
-    defaultValue:
-      memorySizeInKB: 16384
-      mountROSysfs: true
-      volumes: []
-      devices: []
-  timeoutInSeconds:
-    type: "number"
-    required: false
-    defaultValue: 3
-  platforms:
-    type: "object"
-    required: false
-    defaultValue:
-      - os: "*"
-  componentDependencies:
-    type: "object"
-    required: false
-    defaultValue: {}
-  maxQueueSize:
-    type: "number"
-    required: false
-    defaultValue: 1000
-  statusTimeoutInSeconds:
-    type: "number"
-    required: false
-    defaultValue: 60
-  inputPayloadEncodingType:
-    type: "string"
-    required: false
-    defaultValue: "json"
-  containerMode:
-    type: "string"
-    required: false
-    defaultValue: "GreengrassContainer"
